+++ tests/clients/alertrules/test_file_type_mismatch.py
@@ -1,6 +1,6 @@
 import json
 
-from py42.clients.alertrules import FileTypeMismatchClient
+from py42.clients.alertrules.file_type_mismatch import FileTypeMismatchClient
 
 
 class TestFileTypeMisMatchClient(object):
